+++ src/lib/radiobutton.cpp
@@ -43,7 +43,7 @@ RadioButton::RadioButton(const QString& text, QWidget* parent)
 */
 void RadioButton::setReadOnly(bool ro)
 {
-    if ((int)ro != (int)mReadOnly)
+    if (ro != mReadOnly)
     {
         mReadOnly = ro;
         setFocusPolicy(ro ? Qt::NoFocus : mFocusPolicy);
